requirements.txt
playwright
langchain-openai
pydantic
autopep8
python-dotenv